ok, lump it all in here, it seems to work?
